requirements.txt
pyspark
kafka-python
avro-python3